refactor/#66: modify typo
