+++ src/entities/FreeBoard.ts
@@ -1,5 +1,5 @@
 import { FreeBoardHistory } from '@src/entities/FreeBoardHistory';
-import { BooleanTransformer } from '@src/entities/transfomers/boolean.transfomer';
+import { BooleanTransformer } from '@src/entities/transformers/boolean.transformer';
 import {
   Column,
   Entity,
